+++ telegram/error.py
@@ -59,8 +59,7 @@ class TelegramError(Exception):
     """
     Base class for Telegram errors.
 
-    .. seealso:: `Exceptions, Warnings and Logging <https://github.com/\
-    python-telegram-bot/python-telegram-bot/wiki/Exceptions%2C-Warnings-and-Logging>`_
+    .. seealso:: :wiki:`Exceptions, Warnings and Logging <Exceptions%2C-Warnings-and-Logging>`
     """
 
     __slots__ = ("message",)
@@ -149,9 +148,8 @@ class ChatMigrated(TelegramError):
     """
     Raised when the requested group chat migrated to supergroup and has a new chat id.
 
-    .. seealso:: `Storing Bot, User and Chat Related Data <https://github.com/\
-        python-telegram-bot/python-telegram-bot/wiki/Storing-bot%2C-user-and-\
-        chat-related-data>`_,
+    .. seealso::
+        :wiki:`Storing Bot, User and Chat Related Data <Storing-bot%2C-user-and-chat-related-data>`
 
     Args:
         new_chat_id (:obj:`int`): The new chat id of the group.