+++ telegram/ext/_callbackcontext.py
@@ -72,8 +72,7 @@ class CallbackContext(Generic[BT, UD, CD, BD]):
         * :any:`Custom Webhook Bot <examples.customwebhookbot>`
 
     .. seealso:: :attr:`telegram.ext.ContextTypes.DEFAULT_TYPE`,
-        `Job Queue <https://github.com/python-telegram-bot/
-        python-telegram-bot/wiki/Extensions-%E2%80%93-JobQueue>`_
+        :wiki:`Job Queue <Extensions-%E2%80%93-JobQueue>`
 
     Args:
         application (:class:`telegram.ext.Application`): The application associated with this
@@ -146,9 +145,8 @@ class CallbackContext(Generic[BT, UD, CD, BD]):
         """:obj:`ContextTypes.bot_data`: Optional. An object that can be used to keep any data in.
         For each update it will be the same :attr:`ContextTypes.bot_data`. Defaults to :obj:`dict`.
 
-        .. seealso:: `Storing Bot, User and Chat Related Data <https://github.com/\
-            python-telegram-bot/python-telegram-bot/wiki/Storing-bot%2C-user-and-\
-            chat-related-data>`_,
+        .. seealso:: :wiki:`Storing Bot, User and Chat Related Data\
+            <Storing-bot%2C-user-and-chat-related-data>`
         """
         return self.application.bot_data
 
@@ -166,13 +164,11 @@ class CallbackContext(Generic[BT, UD, CD, BD]):
 
         Warning:
             When a group chat migrates to a supergroup, its chat id will change and the
-            ``chat_data`` needs to be transferred. For details see our `wiki page
-            <https://github.com/python-telegram-bot/python-telegram-bot/wiki/
-            Storing-bot,-user-and-chat-related-data#chat-migration>`_.
+            ``chat_data`` needs to be transferred. For details see our
+            :wiki:`wiki page <Storing-bot,-user-and-chat-related-data#chat-migration>`.
 
-        .. seealso:: `Storing Bot, User and Chat Related Data <https://github.com/\
-            python-telegram-bot/python-telegram-bot/wiki/Storing-bot%2C-user-and-\
-            chat-related-data>`_,
+        .. seealso:: :wiki:`Storing Bot, User and Chat Related Data\
+            <Storing-bot%2C-user-and-chat-related-data>`
 
         .. versionchanged:: 20.0
             The chat data is now also present in error handlers if the error is caused by a job.
@@ -193,9 +189,8 @@ class CallbackContext(Generic[BT, UD, CD, BD]):
         For each update from the same user it will be the same :obj:`ContextTypes.user_data`.
         Defaults to :obj:`dict`.
 
-        .. seealso:: `Storing Bot, User and Chat Related Data <https://github.com/\
-            python-telegram-bot/python-telegram-bot/wiki/Storing-bot%2C-user-and\
-            -chat-related-data>`_,
+        .. seealso:: :wiki:`Storing Bot, User and Chat Related Data\
+            <Storing-bot%2C-user-and-chat-related-data>`
 
         .. versionchanged:: 20.0
             The user data is now also present in error handlers if the error is caused by a job.
@@ -244,8 +239,7 @@ class CallbackContext(Generic[BT, UD, CD, BD]):
             Will *not* raise exceptions in case the data is not found in the cache.
             *Will* raise :exc:`KeyError` in case the callback query can not be found in the cache.
 
-        .. seealso:: `Arbitrary callback_data <https://github.com/\
-            python-telegram-bot/python-telegram-bot/wiki/Arbitrary-callback_data>`_
+        .. seealso:: :wiki:`Arbitrary callback_data <Arbitrary-callback_data>`
 
         Args:
             callback_query (:class:`telegram.CallbackQuery`): The callback query.
@@ -387,8 +381,7 @@ class CallbackContext(Generic[BT, UD, CD, BD]):
         :class:`telegram.ext.JobQueue`: The :class:`JobQueue` used by the
             :class:`telegram.ext.Application`.
 
-        .. seealso:: `Job Queue <https://github.com/python-telegram-bot/
-            python-telegram-bot/wiki/Extensions-%E2%80%93-JobQueue>`_
+        .. seealso:: :wiki:`Job Queue <Extensions-%E2%80%93-JobQueue>`
         """
         if self._application._job_queue is None:  # pylint: disable=protected-access
             warn(