+++ telegram/ext/_commandhandler.py
@@ -87,8 +87,7 @@ class CommandHandler(BaseHandler[Update, CCT]):
             be awaited before processing the next handler in
             :meth:`telegram.ext.Application.process_update`. Defaults to :obj:`True`.
 
-            .. seealso:: `Concurrency <https://github.com/\
-                python-telegram-bot/python-telegram-bot/wiki/Concurrency>`_
+            .. seealso:: :wiki:`Concurrency`
 
     Raises:
         :exc:`ValueError`: When the command is too long or has illegal chars.